pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = ""
authors = ["Bileam Scheuvens <36153336+bileamScheuvens@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^3.2.1"
nltk = "^3.8.1"
numpy = "^1.26.2"
pattern = "^3.6.0"
clingo = "^5.6.2"
torch = "^2.3.0"
torchvision = "^0.18.0"
transformers = "^4.36.1"
pillow = "^10.2.0"
pandas = "^2.2.0"
openai = "0.27.8"
huggingface-hub = "^0.20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
